exploration/requirements.txt
astroid==2.3.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pip-autoremove==0.9.1
psycopg2==2.8.3
pylint==2.4.2
PyYAML==5.1.2
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
